Move Exporter to separate module, make module publishable, and update Tools readme
