+++ build.sbt
@@ -173,7 +173,8 @@ lazy val bifrost = project
     toplRpc,
     benchmarking,
     crypto,
-    brambl
+    brambl,
+    tools
   )
 
 lazy val node = project
@@ -201,7 +202,7 @@ lazy val node = project
   .settings(
     IntegrationTest / parallelExecution := false
   )
-  .dependsOn(common % "compile->compile;test->test", toplRpc)
+  .dependsOn(common % "compile->compile;test->test", toplRpc, tools)
   .enablePlugins(BuildInfoPlugin, JavaAppPackaging, DockerPlugin)
 
 lazy val common = project
@@ -303,6 +304,18 @@ lazy val crypto = project
     buildInfoKeys := Seq[BuildInfoKey](name, version, scalaVersion, sbtVersion),
     buildInfoPackage := "co.topl.buildinfo.crypto",
     libraryDependencies ++= Dependencies.crypto,
+  )
+
+lazy val tools = project
+  .in(file("tools"))
+  .enablePlugins(BuildInfoPlugin)
+  .settings(
+    name := "tools",
+    commonSettings,
+    publishSettings,
+    buildInfoKeys := Seq[BuildInfoKey](name, version, scalaVersion, sbtVersion),
+    buildInfoPackage := "co.topl.buildinfo.tools",
+    libraryDependencies ++= Dependencies.tools
   )
 
 addCommandAlias("checkPR", "; scalafixAll --check; scalafmtCheckAll; test")
