--- node/src/main/scala/co/topl/tools/Tools.md
@@ -1,9 +1,0 @@
-## Tools
-
-### Exporter
-
-#### Purpose
-The exporter tool is used to take data structured in a Node LevelDB database
-and export it to another format. This is useful for data inspection
-and queries that can be run against blockchain data if the data is formatted 
-differently or in a flatter structure.
