+++ tools/Tools.md
@@ -1,0 +1,18 @@
+# Tools
+
+## Exporter
+
+### Purpose
+The exporter tool is used to take data structured in a Bifrost Node LevelDB database and export it to another format. This is useful for data inspection and queries that can be run against blockchain data if the data is formatted differently or in a flatter structure.
+
+### Usage
+
+Running the export process requires a Node database with blocks and state that already has all of the data you want to export. For example, if you want to export the first 100,000 blocks from the mainnet database, you need a `.bifrost/toplnet` folder generated by a node that has synced up to or past the requisite block height.
+
+To then export the blocks to a local MongoDB database:
+
+`sbt "node/run export --network toplnet --start 1 --end 100000 --uri mongodb://localhost:27017`
+
+or from a JAR:
+
+`java -jar bifrost-<version>.jar export --network toplnet --start 1 --end 100000 --uri mongodb://localhost:27017`
